requirements.txt
asgiref==3.6.0
Django==4.1.5
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytils==0.4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
